Add logged in and logged out state to header
